explicit cast unnecessary

+++ ArabicToChineseTryThree.java
@@ -186,7 +186,11 @@ public class ArabicToChineseTryThree {
 
     //Now we are beginning to convert the decimal part.
     for(int i=0;i<decimalPart.length();i++) {
-      int charToInt = (int) decimalPart.charAt(i) - 48;
+      /*
+       * I think explicit cast is unnecessary here,because java has a type promotion mechanism
+       * and the decimalPart.charAt(i)(which is of type char) will be promoted to int.
+      */
+      int charToInt =  decimalPart.charAt(i) - 48;
       int correspondingIndex = i;
       if(decimalPart.charAt(i) == '0') {
         numOfConcecutiveZero++;
